feat(BaseHead): Add default props for title and description
This commit adds default props for the `title` and `description` variables in the `BaseHead.astro` component. The default values are set to "MoltenObsidian" for the title and a description related to Obsidian-flavoured Markdown parsing for Blazor with Vault mapping support. Additionally, it includes new links to different sizes of icons and sets the og:image property to use the android-chrome-192x192.png image.