feat(vaults): Add event for vault updates + FS Vault implementation
This commit adds a new event handler `VaultUpdate` to the `IVault` interface. The event is raised when a file, folder, or note is added, removed, updated, or moved within the vault. The event handler is defined as a delegate `VaultUpdateEventHandler`. Additionally, a new class `VaultUpdateEventArgs` is introduced to represent the event arguments for when a vault is updated. It contains properties for the type of update that occurred (`Add`, `Remove`, `Update`, or `Move`) and the entity that was updated (folder, file, or note). This commit also includes changes in the implementation of the `FileSystemVault` class to invoke the `VaultUpdate` event appropriately when items are created, renamed, changed, or deleted within the vault.