feat(vaults/http): Add event for Vault updates
This commit adds a new event `VaultUpdate` to the `HttpRemoteVault` class in the `Vaults.Http` namespace. The event is triggered when there are updates to the vault, although currently there is no change detection implemented within the HTTP Vault. This event can be used in the future to handle vault update notifications in client applications.