feat(vaults/ftp): Add event for vault updates
This commit adds a new event `VaultUpdate` to the `FtpRemoteVault` class, which will be triggered when there are updates to the vault. The event is currently not implemented with change detection. Additionally, the commit includes a new method `FromManifest` that builds a new FTP remote vault from a given manifest and FTP client.

Changes:
- Added `VaultUpdate` event to `FtpRemoteVault`
- Implemented `FromManifest` method in `FtpRemoteVault`

Note: There is no specific mention of files or names in the code changes.