feat(tool): add `--generate-manifest` flag to SSG command
This commit adds a new feature to the SSG command by introducing the `--generate-manifest` flag. When this flag is used, a manifest file will be generated for the local vault if it is missing. The manifest file contains information about all the files in the vault.

The changes include:
- Added a new boolean property `GenerateManifest` to the `GenerateStaticSiteCommandSettings` class.
- Modified the `ExecuteAsync` method in the `GenerateStaticSite` class to check if the `GenerateManifest` flag is set and generate a manifest using the `GenerateManifestAsync` method from the `GenerateManifestCommand`.
- Created a new internal static method called `WriteStaticFilesAsync`, which writes static files to disk.
- Updated existing code to use this new method instead of duplicating code.

These changes enhance the functionality of the SSG command by providing an option to generate a manifest file for better organization and management of files in the vault.