docs: Update flags list
Updated the flags list in two files to reflect recent changes in the codebase. Added new flag `--generate-manifest` to the "Static Site Generation" file and added a note about it being similar to the [[Vault Manifests]] feature. Also, added a note about the `--watch` flag implicitly enabling the `--force` flag in both files.